requirements.txt
fastapi
uvicorn
requests
python-dotenv
collections-extended
langdetect